sie-study-app/pyproject.toml
[project]
name = "sie-study-app"
version = "0.1.0"
description = "SIE exam study app with AI-powered practice questions and progress tracking"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.43.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]
